+++ game/factions/faction.py
@@ -330,10 +330,10 @@ class Faction:
             self.remove_aircraft("F-15D")
         if not mod_settings.f_16_idf:
             self.remove_aircraft("F-16I")
-            self.remove_aircraft("F_16D_52")
-            self.remove_aircraft("F_16D_50")
-            self.remove_aircraft("F_16D_50_NS")
-            self.remove_aircraft("F_16D_52_NS")
+            self.remove_aircraft("F-16D_52")
+            self.remove_aircraft("F-16D_50")
+            self.remove_aircraft("F-16D_50_NS")
+            self.remove_aircraft("F-16D_52_NS")
             self.remove_aircraft("F-16D_Barak_30")
             self.remove_aircraft("F-16D_Barak_40")
         else: